Exercício 2 - Praticando TDD/FiltroDeFaturas/src/main/java/fatura/Fatura.java
package fatura;

import java.util.Date;

public class Fatura {

    private String codigo;
    private double valor;
    private Date data;
    private Cliente cliente;

    public Fatura(String codigo, double valor, Date data, Cliente cliente) {
        this.codigo = codigo;
        this.valor = valor;
        this.data = data;
        this.cliente = cliente;
    }

    public String getCodigo() {
        return codigo;
    }

    public void setCodigo(String codigo) {
        this.codigo = codigo;
    }

    public double getValor() {
        return valor;
    }

    public void setValor(double valor) {
        this.valor = valor;
    }

    public Date getData() {
        return data;
    }

    public void setData(Date data) {
        this.data = data;
    }

    public Cliente getCliente() {
        return cliente;
    }

    public void setCliente(Cliente cliente) {
        this.cliente = cliente;
    }
}
